+++ src/main/java/com/javatar/uni2/repository/StudentDegreeRepository.java
@@ -3,5 +3,5 @@ package com.javatar.uni2.repository;
 import com.javatar.uni2.model.StudentDegree;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentDegreeRepository extends JpaRepository<StudentDegree, Integer> {
+public interface StudentDegreeRepository extends JpaRepository<StudentDegree, Long> {
 }
